materra-django/requirements.txt
channels==4.2.0
channels_redis==4.2.1
daphne==4.1.2
django==5.1.6
djangorestframework==3.15.2
psycopg2-binary==2.9.10
